# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Khojsansar API",
        "description": "Khojsansar App API Documentation",
        "contact": {
            "email": "trishachhetri27@gmail.com"
        },
        "license": {
            "name": "Apache 2.4.58",
            "url": "https://www.apache.org/licenses/LICENSE-2.4.58.html"
        },
        "version": "3.0"
    },
    "paths": {
...
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurant Details"
                ],
                "summary": "Get list of restaurants",
                "operationId": "d2d43182e1d3ee25a848555dbbed79e1",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RestaurantDetail"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/restaurant/{id}": {
            "get": {
                "tags": [
                    "Restaurant Details"
                ],
                "summary": "Get restaurant details by ID",
                "operationId": "fcad4f5c470f4997d46545eb1a4ebcd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/getRestaurantByMunicipalities": {
            "post": {
                "tags": [
                    "Restaurants Detail By Location"
                ],
                "summary": "Get list of restaurants by municipalities with pagination",
                "operationId": "dbe5d36052731d00115c4d44b4e04279",
                "parameters": [
                    {
                        "name": "municipality",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RestaurantDetail"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/getRestaurantsByLatLng": {
            "post": {
                "tags": [
                    "Restaurants Detail By Location"
                ],
                "summary": "Get restaurants by latitude and longitude",
                "operationId": "21b6495610824c92cfebbc664eaa6eb1",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RestaurantDetail"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/restaurantsrate": {
            "get": {
                "tags": [
                    "Restaurants Display as per rating"
                ],
                "summary": "Get list of restaurants sorted by rating",
                "operationId": "8b304bbfb8fe011bc1470b1997035fb6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RestaurantDetail"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RestaurantDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ani's Restaurant"
                    },
                    "restaurant_type": {
                        "type": "string",
                        "example": "Chinese Cuisine"
                    },
                    "coverimage": {
                        "type": "string",
                        "example": "coverimage.jpg"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "review_count": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Ani's Restaurant",
                    "restaurant_type": "Chinese Cuisine",
                    "coverimage": "coverimage.jpg",
                    "rating": 4.5,
                    "review_count": 120
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}